Add LoRA utility functions for fuzzy matching and path resolution
- Implemented functions to retrieve available LoRA files and normalize path separators.
- Added fuzzy matching logic to handle renamed LoRAs and WAN-related tokens.
- Created functions to strip file extensions, normalize names, and resolve LoRA paths.
- Enhanced compatibility with ComfyUI's folder structure for loading LoRAs.
